requirements.txt
boto3==1.9.145
Django==2.2
django-cors-headers==3.0.0
django-extensions==2.1.6
django-filter==2.1.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
honcho==1.0.1
Pillow==6.0.0
psycopg2-binary==2.8.2
PyJWT==1.7.1
pytz==2019.1
rest-framework-apicontrol==0.7.7
rest-framework-auth0==0.5.3
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.2
